requirements.txt
boto3
unittest2
python-dotenv
redis
flask 
flask-socketio